Super-app step 2: shell app-switcher + per-app routing, sidebars, recents
- Routes: /apps home (/ redirects), /apps/<tag>/<name> builder views,
  /explorer launcher + /explorer/view|embed/<path>, /learn, and plain
  /preferences /templates /mounts (underscore sentinels retired; legacy
  /view + /embed URLs rewritten at boot).
- Sidebar becomes a thin context switcher (shell / explorer / builder);
  bookmarks tree moves to apps/explorer, recents split into independent
  explorer (recents.json) and builder (/api/apps/recents) stores.
- Builder views pin modes to claude_split/versions, drop breadcrumbs;
  learn renders chrome-free; old shell Home page removed.
- /apps hub: drop page header and New-app slide-over (hero composer is
  the creation path), hero shows Fused mark + "Fused App"; brand title
  names the active context.
