+++ ARCHITECTURE.md
@@ -51,7 +51,7 @@ fused-render/
 │   │   ├── session.py          # per-file sidecar helpers, router: /api/session GET/PUT
 │   │   ├── ai.py               # fused.ai: Claude CLI binary resolution, _AiSession, router: /api/ai
 │   │   └── routers/            # route-wiring-only modules (no standalone reusable logic)
-│   │       ├── shell.py        # "/", "/view/{path}", "/embed/{path}" (serves the built React shell)
+│   │       ├── shell.py        # "/", "/apps...", "/explorer...", settings pages + legacy "/view","/embed" (serves the built React shell)
 │   │       ├── config.py       # /api/mounts/health, /api/config, /api/desktop/shutdown
 │   │       ├── fs_read.py      # /api/fs/stat|conditions|list|walk|raw|events|reveal
 │   │       ├── render.py       # /render
@@ -105,8 +105,8 @@ fused-render [--start-dir DIR] [--port N] [--no-browser]
 
 All paths in query strings are **absolute filesystem paths**. Server never scopes/rejects by location (v1 has no security layer — deliberate, see SPEC §9). Errors return `{"error": "<message>"}` with 4xx status. Every response carries `Cache-Control: no-cache` (middleware) — app code changes between restarts and user files change on disk; stale cached shell/runtime JS produced half-old UIs during development. The two mutating/executing POSTs (`/api/run`, `/api/fs/write`) require an `X-Fused: 1` header (missing/wrong → 403); it forces a CORS preflight so a foreign page can't fire them blind. Not auth — D3 stands (see DECISIONS.md D36).
 
-### `GET /` and `GET /view/{path:path}` → shell.html
-Same static shell for both; shell JS reads `location.pathname` to route. `/view/Users/you/data` means fs path `/Users/you/data` (strip `/view/`, prepend `/`).
+### `GET /` and `GET /explorer/view/{path:path}` → shell.html
+Same static shell for both; shell JS reads `location.pathname` to route. `/explorer/view/Users/you/data` means fs path `/Users/you/data` (strip `/explorer/view/`, prepend `/`). Legacy `/view/` and `/embed/` URLs are rewritten in place at boot (router.ts).
 
 ### `GET /api/fs/stat?path=<abs>`
 ```json
@@ -212,7 +212,7 @@ Behavior:
 - **Theme (SPEC §30, D134):** the runtime is also how the appearance reaches a view document, since a React re-render must never touch a live iframe. It runs at the very top of the IIFE — `/render` injects it right after `<head>`, parser-blocking, so it executes before the document's own stylesheet is parsed and there is no flash — and, **only if `<html>` carries `data-fused-theme`**, resolves the theme and writes `data-theme` on that same element. It resolves *itself* (same-origin `localStorage["fused-render:theme"]`, falling back to this document's `matchMedia`) rather than being pushed from the shell, and follows changes via the `storage` event (another window's override) plus its own `matchMedia` change (a System-mode OS flip). Nothing is ever written into a document without the opt-in, so user-authored `.html` views get no theme signal at all. The opt-in must be an `<html>` attribute, not a `<meta>`, precisely because of the injection point.
 
 Top-level `path` handling in shell URL vs iframe URL:
-- Shell URL: `/view/<fs-path>?freq=2.4` — params live here (source of truth, PR-1).
+- Shell URL: `/explorer/view/<fs-path>?freq=2.4` — params live here (source of truth, PR-1).
 - Iframe URL: `/render?path=<abs html path>` — no user params needed on it; runtime reads/writes **parent's** query string.
 - Standalone fallback (`parent === window`): read/write own URL's query, skipping `path`.
 
@@ -221,7 +221,7 @@ Top-level `path` handling in shell URL vs iframe URL:
 ## 6. Shell (`frontend/` → `static/shell-dist/`)
 
 SPA, React 18 + Vite (D52/D53; TypeScript, strict). `src/` is layered into four aliased areas — the super-app structure: **`@platform`** (shared foundations: `platform/lib/` non-React modules ported ~verbatim from the vanilla shell — router/api/format/bookmarks/layout-codec, same contracts as before — plus `platform/ui/` shared components and `platform/cloud/` deploy/clone UI used by more than one app), **`@apps/explorer`** (Listing/Preview/Panel/Tabs/Breadcrumb + fs-actions/fs-clipboard), **`@apps/builder`** (Apps hub, NewAppPanel, AppPreviewCard), **`@apps/learn`** (the doorway into the mounted learn content — the experience itself is html+py rendered by the engine), and **`@shell`** (App, Home, Sidebar, Preferences/Account/Mounts/Templates — the chrome that composes the apps). Import direction is enforced by `frontend/scripts/check-boundaries.mjs` (runs first in `npm run build`): platform imports only platform; an app imports only platform and itself — never shell, never another app; shell may import anything. Within that: the router never imports UI (it dispatches a `fused:navigate` event; `platform/lib/hooks.ts` turns it plus `popstate` into a **nav epoch** that keys — i.e. remounts — the active view, the React equivalent of the vanilla per-route DOM rebuild), and the bookmark store never touches the DOM (mutations signal via `notifyBookmarksChanged()`). `Breadcrumb.tsx` may import `Panel.tsx` (`panelUrl`) — both explorer — and shell's `Sidebar.tsx` may import `@apps/explorer/Tabs` (`composeFolderTabsUrl`), since no view imports back — no cycles. The history replaceState/pushState wrapping (→ `fused:urlchange`) lives in `main.tsx` and is load-bearing for the iframe runtimes (D46), not just for the shell's own re-renders; chrome (bookmark buttons, active highlight) re-renders on a **url version** signal that also counts `fused:urlchange`, without remounting views. Layout-mode iframes freeze their `src` at mount — React never rewrites it (a src write reloads an iframe); pane crumb clicks write it imperatively via a ref, and tab frames render as a flat keyed list that only appends/removes (never re-parents/reorders). Routing from `location.pathname`:
-- `/` → redirect (replaceState) to `/view/<start-dir>` (start dir from `GET /api/config` → `{"start_dir": "/Users/you", "home": …}` — `source_template` was dropped with the html sentinel modes (D62); the code-view path arrives via `stat.templates` like everything else, and the shell `Config` type dropped it too).
+- `/` → redirect (replaceState) to `/apps` (the app home; super-app step 2) (start dir from `GET /api/config` → `{"start_dir": "/Users/you", "home": …}` — `source_template` was dropped with the html sentinel modes (D62); the code-view path arrives via `stat.templates` like everything else, and the shell `Config` type dropped it too).
 - `/view/<path>` → `stat` it:
   - a target with a non-empty `stat.templates` → preview view — **including a directory** (every directory resolves at least the universal `/` key → `["_listing"]`, SPEC PT-13/D81; the built-in listing is the `_listing` sentinel mode)
   - **dir** with an empty `templates` (a `null` binding disabled it) → listing view — the shell's safety net (a folder must always render something)
