+++ frontend/src/platform/lib/hooks.ts
@@ -17,7 +17,7 @@ import { NAV_EVENT } from "@platform/lib/router";
 import { createCloseDeferrer } from "@platform/lib/exit-animation";
 import { getConfig } from "@platform/lib/api";
 
-function useEventCounter(events: readonly string[]): number {
+export function useEventCounter(events: readonly string[]): number {
   const [n, setN] = useState(0);
   useEffect(() => {
     const bump = () => setN((v) => v + 1);
@@ -65,20 +65,6 @@ export function notifyArmedChanged(): void {
 
 export function useArmedVersion(): number {
   return useEventCounter([ARMED_EVENT]);
-}
-
-// Recents store change signal — same pattern as the bookmarks event. The
-// recents store (lib/recents.ts) dispatches it itself after every cache
-// advance (its mutations are triggered by tracking, not by UI clicks alone,
-// so the store owns the notify rather than each call site).
-const RECENTS_EVENT = "fused:recents";
-
-export function notifyRecentsChanged(): void {
-  window.dispatchEvent(new Event(RECENTS_EVENT));
-}
-
-export function useRecentsVersion(): number {
-  return useEventCounter([RECENTS_EVENT]);
 }
 
 // Run `cb` when the tab regains focus or becomes visible again — the app's
