+++ frontend/src/platform/lib/router.ts
@@ -1,18 +1,67 @@
-// fs-path <-> /view/ URL codec + navigation. UI-free. The vanilla shell
-// registered a route() handler here; the React shell instead listens for the
-// "fused:navigate" event (useNavEpoch in lib/hooks.ts) — navigate/navigateUrl
-// dispatch it after pushState, popstate is subscribed alongside it.
-export const VIEW_PREFIX = "/view/";
+// fs-path <-> /explorer/view/ URL codec + navigation. UI-free. The vanilla
+// shell registered a route() handler here; the React shell instead listens for
+// the "fused:navigate" event (useNavEpoch in lib/hooks.ts) — navigate/
+// navigateUrl dispatch it after pushState, popstate is subscribed alongside it.
+export const VIEW_PREFIX = "/explorer/view/";
 
 // Embed = chrome-free variant of view (same shell, same routing, just no
 // sidebar/breadcrumb/preview-header). The mode is fixed at page load: both
 // prefixes are served by full page loads, so it can't change without one.
-export const EMBED_PREFIX = "/embed/";
+export const EMBED_PREFIX = "/explorer/embed/";
+
+// Pre-rename URL shapes (old bookmarks/recents entries, .bookmark files,
+// external embed links) rewritten in place at module init — before IS_EMBED
+// is computed, so a legacy /embed/ load still comes up in embed mode. Settings
+// sentinels became plain routes at the same time.
+const LEGACY_SENTINELS: Record<string, string> = {
+  "/view/_home": "/apps",
+  "/view/_prefs": "/preferences",
+  "/view/_templates": "/templates",
+  "/view/_mounts": "/mounts",
+  "/view/_account": "/preferences",
+};
+(function rewriteLegacyPath(): void {
+  const p = location.pathname;
+  const strippedQuery = p === "/view/_account" ? "?tab=account" : location.search;
+  const mapped = LEGACY_SENTINELS[p];
+  let next: string | null = null;
+  if (mapped) next = mapped + strippedQuery;
+  else if (p.startsWith("/view/")) next = "/explorer" + p + location.search;
+  else if (p.startsWith("/embed/")) next = "/explorer" + p + location.search;
+  if (next) history.replaceState(history.state, "", next);
+})();
+
 export const IS_EMBED =
-  location.pathname.startsWith(EMBED_PREFIX) || location.pathname === "/embed";
+  location.pathname.startsWith(EMBED_PREFIX) ||
+  location.pathname === "/explorer/embed";
 // URL prefix for this page's mode. Keeps refresh, in-listing navigation, and
 // param sync (iframe runtime's history.replaceState) inside the active prefix.
 const PREFIX = IS_EMBED ? EMBED_PREFIX : VIEW_PREFIX;
+
+// App-builder namespace: /apps/<tag>/<name> — pretty URLs for app folders,
+// which live exactly two levels under the workspace (<fused_dir>/<tag>/<name>,
+// see fused_render/server/routers/apps.py). Pure codec against fused_dir; no
+// server lookup. /apps itself (no segments) is the apps homepage, not an app.
+export const APP_ROUTE_PREFIX = "/apps/";
+
+export function appUrlForFsPath(fsPath: string, fusedDir: string): string | null {
+  const root = fusedDir.replace(/\/+$/, "");
+  if (!fsPath.startsWith(root + "/")) return null;
+  const segs = fsPath.slice(root.length + 1).split("/").filter((s) => s.length > 0);
+  if (segs.length !== 2) return null;
+  return APP_ROUTE_PREFIX + segs.map(encodeURIComponent).join("/");
+}
+
+export function fsPathFromAppRoute(pathname: string, fusedDir: string): string | null {
+  if (!pathname.startsWith(APP_ROUTE_PREFIX)) return null;
+  const segs = pathname
+    .slice(APP_ROUTE_PREFIX.length)
+    .split("/")
+    .filter((s) => s.length > 0)
+    .map(decodeURIComponent);
+  if (segs.length !== 2) return null;
+  return fusedDir.replace(/\/+$/, "") + "/" + segs.join("/");
+}
 
 export const NAV_EVENT = "fused:navigate";
