+++ fused_render/templates/graph/template.html
@@ -146,9 +146,9 @@
     function shellPrefix() {
       try {
         const p = window.top.location.pathname;
-        if (p.startsWith("/embed/") || p === "/embed") return "/embed/";
+        if (p.startsWith("/explorer/embed/") || p === "/explorer/embed") return "/explorer/embed/";
       } catch (e) { /* cross-origin — assume the normal shell */ }
-      return "/view/";
+      return "/explorer/view/";
     }
     function navigateShell(fsPath) {
       const norm = /^[A-Za-z]:[\\/]/.test(fsPath) ? fsPath.replace(/\\/g, "/") : fsPath;
