+++ fused_render/templates/notebook/template.html
@@ -1786,9 +1786,9 @@ Z / Shift+Z — undo / redo · S — save">⌨</span>
     // session keeps routing under /embed/, never teleported to /view/.
     try {
       const p = window.top.location.pathname;
-      if (p.startsWith("/embed/") || p === "/embed") return "/embed/";
+      if (p.startsWith("/explorer/embed/") || p === "/explorer/embed") return "/explorer/embed/";
     } catch (e) { /* cross-origin — assume the normal shell */ }
-    return "/view/";
+    return "/explorer/view/";
   }
   function viewUrlFor(path) {
     // mirror the shell's /view codec (frontend/src/lib/router.ts): forward
