+++ tests/test_view_url_codec.py
@@ -17,15 +17,15 @@ from fused_render._view_url_codec import (
 
 
 def test_drive_letter_path():
-    assert view_url(8000, "C:\\Users\\x") == "http://127.0.0.1:8000/view/C%3A/Users/x"
+    assert view_url(8000, "C:\\Users\\x") == "http://127.0.0.1:8000/explorer/view/C%3A/Users/x"
 
 
 def test_drive_letter_path_unicode_and_spaces():
     # Mirrors the Rust supervisor test path C:\data\résumé.xlsx.
-    assert view_url_path("C:\\data\\résumé.xlsx") == "/view/C%3A/data/r%C3%A9sum%C3%A9.xlsx"
+    assert view_url_path("C:\\data\\résumé.xlsx") == "/explorer/view/C%3A/data/r%C3%A9sum%C3%A9.xlsx"
     assert (
         view_url_path("C:\\data\\my résumé v2.xlsx")
-        == "/view/C%3A/data/my%20r%C3%A9sum%C3%A9%20v2.xlsx"
+        == "/explorer/view/C%3A/data/my%20r%C3%A9sum%C3%A9%20v2.xlsx"
     )
 
 
@@ -37,25 +37,25 @@ def test_unc_path_stays_one_segment():
     # router.ts urlForFsPath only normalizes drive-letter paths, so a UNC path
     # is one percent-encoded segment, backslashes and all.
     assert view_url(8000, "\\\\server\\share\\file") == (
-        "http://127.0.0.1:8000/view/%5C%5Cserver%5Cshare%5Cfile"
+        "http://127.0.0.1:8000/explorer/view/%5C%5Cserver%5Cshare%5Cfile"
     )
 
 
 def test_posix_backslash_filename_untouched():
     # deeplink runs this codec on POSIX servers too: a backslash there is a
     # legal filename character and must not be treated as a separator.
-    assert view_url_path("/home/user/back\\slash.txt") == "/view/home/user/back%5Cslash.txt"
+    assert view_url_path("/home/user/back\\slash.txt") == "/explorer/view/home/user/back%5Cslash.txt"
 
 
 def test_bookmark_on_drive_path():
     assert view_url(8000, "C:\\Users\\x\\demo.bookmark") == (
-        "http://127.0.0.1:8000/view/_bookmark?file=C%3A%2FUsers%2Fx%2Fdemo.bookmark"
+        "http://127.0.0.1:8000/explorer/view/_bookmark?file=C%3A%2FUsers%2Fx%2Fdemo.bookmark"
     )
 
 
 def test_bookmark_on_unc_path():
     assert view_url(8000, "\\\\server\\share\\demo.bookmark") == (
-        "http://127.0.0.1:8000/view/_bookmark?file=%5C%5Cserver%5Cshare%5Cdemo.bookmark"
+        "http://127.0.0.1:8000/explorer/view/_bookmark?file=%5C%5Cserver%5Cshare%5Cdemo.bookmark"
     )
 
 
@@ -89,26 +89,26 @@ def test_open_target_file_uri_matches_bare_path():
     assert open_target_path("file:///home/u/a.parquet") == open_target_path(
         "/home/u/a.parquet"
     )
-    assert open_target_path("file:///home/u/a.parquet") == "/view/home/u/a.parquet"
+    assert open_target_path("file:///home/u/a.parquet") == "/explorer/view/home/u/a.parquet"
 
 
 def test_open_target_file_uri_decodes_percent_escapes():
     assert open_target_path("file:///home/u/my%20report.html") == (
-        "/view/home/u/my%20report.html"
+        "/explorer/view/home/u/my%20report.html"
     )
 
 
 def test_open_target_plain_absolute_path():
-    assert open_target_path("/data/report.parquet") == "/view/data/report.parquet"
+    assert open_target_path("/data/report.parquet") == "/explorer/view/data/report.parquet"
 
 
 def test_open_target_folder_path():
-    assert open_target_path("/home/u/project") == "/view/home/u/project"
+    assert open_target_path("/home/u/project") == "/explorer/view/home/u/project"
 
 
 def test_open_target_url_prefixes_host_and_port():
     assert open_target_url(8000, "/data/x.parquet") == (
-        "http://127.0.0.1:8000/view/data/x.parquet"
+        "http://127.0.0.1:8000/explorer/view/data/x.parquet"
     )
     assert open_target_url(8000, "fused-render://open?git=x") == (
         "http://127.0.0.1:8000/clone?src=fused-render%3A%2F%2Fopen%3Fgit%3Dx"
@@ -118,7 +118,7 @@ def test_open_target_url_prefixes_host_and_port():
 def test_open_target_file_uri_localhost_is_local():
     # RFC 8089: an empty authority and "localhost" both mean this machine.
     assert open_target_path("file://localhost/home/u/a.parquet") == (
-        "/view/home/u/a.parquet"
+        "/explorer/view/home/u/a.parquet"
     )
 
 
@@ -133,7 +133,7 @@ def test_open_target_file_uri_with_remote_host_fails_loudly():
 
 def test_open_target_http_url_fails_loudly():
     # Any non-fused-render, non-file scheme has no filesystem path to view;
-    # falling through to view_url_path produced /view/https:/example.com/x.
+    # falling through to view_url_path produced /explorer/view/https:/example.com/x.
     with pytest.raises(OSError):
         open_target_path("https://example.com/x")
     with pytest.raises(OSError):
